requirements.txt
ossapi>=3.1.9
disnake>=2.8.1
psutil>=5.9.5
python-dotenv>=1.0.0
webcolors>=1.13
roblox>=2.0.0
requests>=2.30.0
requests-oauthlib>=1.3.1
asyncio>=3.4.3
async-timeout>=4.0.2
redis>=4.5.5
gdshortener>=1.0.1
topggpy>=1.4.0